// modules/hub/apps/frontend-navegante-app/src/types/motis-api/types.gen.ts
export type Mode = 'WALK' | 'BIKE' | 'RENTAL' | 'CAR' | 'CAR_PARKING' | 'CAR_DROPOFF' | 'ODM' | 'RIDE_SHARING' | 'FLEX' | 'DEBUG_BUS_ROUTE' | 'DEBUG_RAILWAY_ROUTE' | 'DEBUG_FERRY_ROUTE' | 'TRANSIT' | 'TRAM' | 'SUBWAY' | 'FERRY' | 'AIRPLANE' | 'BUS' | 'COACH' | 'RAIL' | 'HIGHSPEED_RAIL' | 'LONG_DISTANCE' | 'NIGHT_RAIL' | 'REGIONAL_FAST_RAIL' | 'REGIONAL_RAIL' | 'SUBURBAN' | 'FUNICULAR' | 'AERIAL_LIFT' | 'OTHER' | 'AREAL_LIFT' | 'METRO' | 'CABLE_CAR';

/**
 * a transit route/line serving a stop
 */
export type Route = {
    routeId: string;
    routeShortName: string;
    routeLongName: string;
    mode: Mode;
    agencyId: string;
    agencyName: string;
    agencyUrl: string;
    routeColor?: string;
    routeTextColor?: string;
    routeType?: number;
};

/**
 * Cause of this alert.
 */
export type AlertCause = 'UNKNOWN_CAUSE' | 'OTHER_CAUSE' | 'TECHNICAL_PROBLEM' | 'STRIKE' | 'DEMONSTRATION' | 'ACCIDENT' | 'HOLIDAY' | 'WEATHER' | 'MAINTENANCE' | 'CONSTRUCTION' | 'POLICE_ACTIVITY' | 'MEDICAL_EMERGENCY' | 'SPECIAL_EVENT';

/**
 * The effect of this problem on the affected entity.
 */
export type AlertEffect = 'NO_SERVICE' | 'REDUCED_SERVICE' | 'SIGNIFICANT_DELAYS' | 'DETOUR' | 'ADDITIONAL_SERVICE' | 'MODIFIED_SERVICE' | 'OTHER_EFFECT' | 'UNKNOWN_EFFECT' | 'STOP_MOVED' | 'NO_EFFECT' | 'ACCESSIBILITY_ISSUE';

/**
 * The severity of the alert.
 */
export type AlertSeverityLevel = 'UNKNOWN_SEVERITY' | 'INFO' | 'WARNING' | 'SEVERE';
// ...
export type Alert = {
    /**
     * Attribute or notice code (e.g. for HRDF or NeTEx)
     */
    code?: string;
    /**
     * Time when the alert should be shown to the user.
     * If missing, the alert will be shown as long as it appears in the feed.
     * If multiple ranges are given, the alert will be shown during all of them.
     *
     */
    communicationPeriod?: Array<TimeRange>;
    /**
     * Time when the services are affected by the disruption mentioned in the alert.
     */
    impactPeriod?: Array<TimeRange>;
    cause?: AlertCause;
    /**
     * Description of the cause of the alert that allows for agency-specific language;
     * more specific than the Cause.
     *
     */
    causeDetail?: string;
    effect?: AlertEffect;
    /**
     * Description of the effect of the alert that allows for agency-specific language;
     * more specific than the Effect.
     *
     */
    effectDetail?: string;
    /**
     * The URL which provides additional information about the alert.
     */
    url?: string;
    /**
     * Header for the alert. This plain-text string will be highlighted, for example in boldface.
     *
     */
    headerText: string;
    /**
     * Description for the alert.
     * This plain-text string will be formatted as the body of the alert (or shown on an explicit "expand" request by the user).
     * The information in the description should add to the information of the header.
     *
     */
    descriptionText: string;
    /**
     * Text containing the alert's header to be used for text-to-speech implementations.
     * This field is the text-to-speech version of header_text.
     * It should contain the same information as headerText but formatted such that it can read as text-to-speech
     * (for example, abbreviations removed, numbers spelled out, etc.)
     *
     */
    ttsHeaderText?: string;
    /**
     * Text containing a description for the alert to be used for text-to-speech implementations.
     * This field is the text-to-speech version of description_text.
     * It should contain the same information as description_text but formatted such that it can be read as text-to-speech
     * (for example, abbreviations removed, numbers spelled out, etc.)
     *
     */
    ttsDescriptionText?: string;
    /**
     * Severity of the alert.
     */
    severityLevel?: AlertSeverityLevel;
    /**
     * String containing an URL linking to an image.
     */
    imageUrl?: string;
    /**
     * IANA media type as to specify the type of image to be displayed. The type must start with "image/"
     *
     */
    imageMediaType?: string;
    /**
     * Text describing the appearance of the linked image in the image field
     * (e.g., in case the image can't be displayed or the user can't see the image for accessibility reasons).
     * See the HTML spec for alt image text.
     *
     */
    imageAlternativeText?: string;
};
// ...
export type Area = {
    /**
     * Name of the area
     */
    name: string;
    /**
     * [OpenStreetMap `admin_level`](https://wiki.openstreetmap.org/wiki/Key:admin_level)
     * of the area
     *
     */
    adminLevel: number;
    /**
     * Whether this area was matched by the input text
     */
    matched: boolean;
    /**
     * Set for the first area after the `default` area that distinguishes areas
     * if the match is ambiguous regarding (`default` area + place name / street [+ house number]).
     *
     */
    unique?: boolean;
    /**
     * Whether this area should be displayed as default area (area with admin level closest 7)
     */
    default?: boolean;
};

/**
 * Matched token range (from index, length)
 */
export type Token = [
    number,
    number
];

/**
 * location type
 */
export type LocationType = 'ADDRESS' | 'PLACE' | 'STOP';

/**
 * GeoCoding match
 */
export type Match = {
    type: LocationType;
    /**
     * Experimental. Type categories might be adjusted.
     *
     * For OSM stop locations: the amenity type based on
     * https://wiki.openstreetmap.org/wiki/OpenStreetMap_Carto/Symbols
     *
     */
    category?: string;
    /**
     * list of non-overlapping tokens that were matched
     */
    tokens: Array<Token>;
    /**
     * name of the location (transit stop / PoI / address)
     */
    name: string;
    /**
     * unique ID of the location
     */
    id: string;
    /**
     * latitude
     */
    lat: number;
    /**
     * longitude
     */
    lon: number;
    /**
     * level according to OpenStreetMap
     * (at the moment only for public transport)
     *
     */
    level?: number;
    /**
     * street name
     */
    street?: string;
    /**
     * house number
     */
    houseNumber?: string;
    /**
     * ISO3166-1 country code from OpenStreetMap
     */
    country?: string;
    /**
     * zip code
     */
    zip?: string;
    /**
     * timezone name (e.g. "Europe/Berlin")
     */
    tz?: string;
    /**
     * list of areas
     */
    areas: Array<Area>;
    /**
     * score according to the internal scoring system (the scoring algorithm might change in the future)
     */
    score: number;
    /**
     * available transport modes for stops
     */
    modes?: Array<Mode>;
    /**
     * importance of a stop, normalized from [0, 1]
     */
    importance?: number;
};
// ...
export type VertexType = 'NORMAL' | 'BIKESHARE' | 'TRANSIT';

/**
 * - `NORMAL` - entry/exit is possible normally
 * - `NOT_ALLOWED` - entry/exit is not allowed
 *
 */
export type PickupDropoffType = 'NORMAL' | 'NOT_ALLOWED';

export type Place = {
    /**
     * name of the transit stop / PoI / address
     */
    name: string;
    /**
     * The ID of the stop. This is often something that users don't care about.
     */
    stopId?: string;
    /**
     * If it's not a root stop, this field contains the `stopId` of the parent stop.
     */
    parentId?: string;
    /**
     * The importance of the stop between 0-1.
     */
    importance?: number;
    /**
     * latitude
     */
    lat: number;
    /**
     * longitude
     */
    lon: number;
    /**
     * level according to OpenStreetMap
     * If no level is given, the field will be unset.
     *
     * For older versions (v1-v5), this field is mandatory and therefore set to 0.
     * Affected endpoints: plan, trip, stoptimes, one-to-all, map/stops, map/trips
     *
     */
    level?: number;
    /**
     * timezone name (e.g. "Europe/Berlin")
     */
    tz?: string;
    /**
     * arrival time
     */
    arrival?: string;
    /**
     * departure time
     */
    departure?: string;
    /**
     * scheduled arrival time
     */
    scheduledArrival?: string;
    /**
     * scheduled departure time
     */
    scheduledDeparture?: string;
    /**
     * scheduled track from the static schedule timetable dataset
     */
    scheduledTrack?: string;
    /**
     * The current track/platform information, updated with real-time updates if available.
     * Can be missing if neither real-time updates nor the schedule timetable contains track information.
     *
     */
    track?: string;
    /**
     * Short, abbreviated identifier of the stop intended for users
     * (e.g. printed on signage, used in SMS / passenger info systems).
     * Comes from the GTFS `stop_code` field.
     *
     */
    stopCode?: string;
    /**
     * description of the location that provides more detailed information
     */
    description?: string;
    vertexType?: VertexType;
    /**
     * Type of pickup. It could be disallowed due to schedule, skipped stops or cancellations.
     */
    pickupType?: PickupDropoffType;
    /**
     * Type of dropoff. It could be disallowed due to schedule, skipped stops or cancellations.
     */
    dropoffType?: PickupDropoffType;
    /**
     * Whether this stop is cancelled due to the realtime situation.
     */
    cancelled?: boolean;
    /**
     * Alerts for this stop.
     */
    alerts?: Array<Alert>;
    /**
     * for `FLEX` transports, the flex location area or location group name
     */
    flex?: string;
    /**
     * for `FLEX` transports, the flex location area ID or location group ID
     */
    flexId?: string;
    /**
     * Time that on-demand service becomes available
     */
    flexStartPickupDropOffWindow?: string;
    /**
     * Time that on-demand service ends
     */
    flexEndPickupDropOffWindow?: string;
    /**
     * available transport modes for stops
     */
    modes?: Array<Mode>;
};
// ...
export type FareTransferRule = 'A_AB' | 'A_AB_B' | 'AB';
// ...
export type FareTransfer = {
    rule?: FareTransferRule;
    transferProducts?: Array<FareProduct>;
    /**
     * Lists all valid fare products for the effective fare legs.
     * This is an `array<array<FareProduct>>` where the inner array
     * lists all possible fare products that would cover this effective fare leg.
     * Each "effective fare leg" can have multiple options for adult/child/weekly/monthly/day/one-way tickets etc.
     * You can see the outer array as AND (you need one ticket for each effective fare leg (`A_AB_B`), the first effective fare leg (`A_AB`) or no fare leg at all but only the transfer product (`AB`)
     * and the inner array as OR (you can choose which ticket to buy)
     *
     */
    effectiveFareLegProducts: Array<Array<Array<FareProduct>>>;
};
// ...
export type Itinerary = {
    /**
     * journey duration in seconds
     */
    duration: number;
    /**
     * journey departure time
     */
    startTime: string;
    /**
     * journey arrival time
     */
    endTime: string;
    /**
     * The number of transfers this trip has.
     */
    transfers: number;
    /**
     * Experimental (format might change). Opaque itinerary identifier.
     * Pass it as `itineraryId` to `refresh-itinerary` endpoint for reconstruction using the new schedule/realtime data.
     *
     */
    id: string;
    /**
     * Journey legs
     */
    legs: Array<Leg>;
    /**
     * Fare information
     */
    fareTransfers?: Array<FareTransfer>;
};
// ...
export type PlanResponse = ({
    /**
     * the routing query
     */
    requestParameters: {
        [key: string]: (string);
    };
    /**
     * debug statistics
     */
    debugOutput: {
        [key: string]: (number);
    };
    from: Place;
    to: Place;
    /**
     * Direct trips by `WALK`, `BIKE`, `CAR`, etc. without time-dependency.
     * The starting time (`arriveBy=false`) / arrival time (`arriveBy=true`) is always the queried `time` parameter (set to \"now\" if not set).
     * But all `direct` connections are meant to be independent of absolute times.
     *
     */
    direct: Array<Itinerary>;
    /**
     * list of itineraries
     */
    itineraries: Array<Itinerary>;
    /**
     * Use the cursor to get the previous page of results. Insert the cursor into the request and post it to get the previous page.
     * The previous page is a set of itineraries departing BEFORE the first itinerary in the result for a depart after search. When using the default sort order the previous set of itineraries is inserted before the current result.
     *
     */
    previousPageCursor: string;
    /**
     * Use the cursor to get the next page of results. Insert the cursor into the request and post it to get the next page.
     * The next page is a set of itineraries departing AFTER the last itinerary in this result.
     *
     */
    nextPageCursor: string;
});
// ...
export type GeocodeResponse = (Array<Match>);